feat(auth): Реализован полный цикл сброса пароля
Добавлена новая, безопасная логика восстановления пароля через email и телефон.

- Созданы 3 новых эндпоинта: /request, /verify-phone и /reset.
- Реализована защита от перебора (brute-force) для кодов верификации.
- Добавлена защита от спама (rate-limiting) при запросе кода.
- 'Магические числа' (TTL, кол-во попыток) вынесены в конфиг.
- В ответ на логин теперь сразу приходят права пользователя.

+++ internal/entities/priority-entity.go
@@ -4,10 +4,8 @@ import (
 	"request-system/pkg/types"
 )
 
-
-
 type Priority struct {
-	ID        int    `json:"id"`
+	ID        uint64 `json:"id"`
 	Icon      string `json:"icon"`
 	Name      string `json:"name"`
 	Rate      int    `json:"rate"`
